Revise H-band Pol contrast with new sat spot peak-to-total flux ratio
In gpi_measure_contrast_pol.pro, the sat spot peak-to-total flux
ratio for H-band is now hardcoded to 1/13.5. When comparing
consecutively observed Spec and Pol datasets, the measured ratio of
Spec satspot peak to Pol satspot integrated flux is < the Gaussian
assumption of ~1/11. The mean is closer to 1/13.5 (e.g. HD 106906
data from 20150504), which is more consistent with a 2d Moffat
profile.

Other bands are unchanged-- they still use the Gaussian assumption.

Also added GRIDFAC header keyword to output contrast profile .fits
file to record the star-to-satspot ratio used in calculating the
contrast.